--- gui/modelDataWidget.hpp
@@ -1,24 +1,0 @@
-#ifndef BLE_GUI_MODELDATAWIDGET_H_
-#define BLE_GUI_MODELDATAWIDGET_H_
-
-#include <iostream>
-#include <memory>
-
-#include <QWidget>
-#include <QDoubleSpinBox>
-
-namespace ble_gui
-{
-
-	class ModelDataWidget : public QWidget
-	{
-		Q_OBJECT
-	public:
-		ModelDataWidget(QWidget *parent = nullptr);
-		~ModelDataWidget() {}
-		QDoubleSpinBox *Period;
-	};
-
-} // namespace ble_gui
-
-#endif
